chore: detect and fix circular import dependencies
Resolves #1061

Changes:
- Created src/debug/systemTiming.ts to break circular dependency
- Moved system timing code from debug/index.ts to systemTiming.ts
- Updated debug/overlay.ts to import from systemTiming.ts
- Modified getPerformanceStats to accept entityCount parameter
- Added check:circular script to package.json for validation

Verified with madge - no circular dependencies found.
