+++ package.json
@@ -142,6 +142,7 @@
     "lint:fix": "biome check --write .",
     "format": "biome format --write .",
     "typecheck": "tsc --noEmit",
+    "check:circular": "madge --circular --extensions ts,tsx src/",
     "prepare": "husky",
     "prepublishOnly": "pnpm run build:publish",
     "build:publish": "tsup --no-sourcemap --minify",
@@ -154,17 +155,98 @@
     {
       "path": "dist/index.js",
       "limit": "500 KB",
-      "ignore": ["fs", "path", "os", "util", "child_process", "net", "tty", "stream", "events", "readline", "module", "url", "https", "node:fs", "node:path", "node:os", "node:util", "node:child_process", "node:net", "node:tty", "node:stream", "node:events", "node:readline", "node:module", "node:url", "node:https"]
+      "ignore": [
+        "fs",
+        "path",
+        "os",
+        "util",
+        "child_process",
+        "net",
+        "tty",
+        "stream",
+        "events",
+        "readline",
+        "module",
+        "url",
+        "https",
+        "node:fs",
+        "node:path",
+        "node:os",
+        "node:util",
+        "node:child_process",
+        "node:net",
+        "node:tty",
+        "node:stream",
+        "node:events",
+        "node:readline",
+        "node:module",
+        "node:url",
+        "node:https"
+      ]
     },
     {
       "path": "dist/core/index.js",
       "limit": "100 KB",
-      "ignore": ["fs", "path", "os", "util", "child_process", "net", "tty", "stream", "events", "readline", "module", "url", "https", "node:fs", "node:path", "node:os", "node:util", "node:child_process", "node:net", "node:tty", "node:stream", "node:events", "node:readline", "node:module", "node:url", "node:https"]
+      "ignore": [
+        "fs",
+        "path",
+        "os",
+        "util",
+        "child_process",
+        "net",
+        "tty",
+        "stream",
+        "events",
+        "readline",
+        "module",
+        "url",
+        "https",
+        "node:fs",
+        "node:path",
+        "node:os",
+        "node:util",
+        "node:child_process",
+        "node:net",
+        "node:tty",
+        "node:stream",
+        "node:events",
+        "node:readline",
+        "node:module",
+        "node:url",
+        "node:https"
+      ]
     },
     {
       "path": "dist/components/index.js",
       "limit": "200 KB",
-      "ignore": ["fs", "path", "os", "util", "child_process", "net", "tty", "stream", "events", "readline", "module", "url", "https", "node:fs", "node:path", "node:os", "node:util", "node:child_process", "node:net", "node:tty", "node:stream", "node:events", "node:readline", "node:module", "node:url", "node:https"]
+      "ignore": [
+        "fs",
+        "path",
+        "os",
+        "util",
+        "child_process",
+        "net",
+        "tty",
+        "stream",
+        "events",
+        "readline",
+        "module",
+        "url",
+        "https",
+        "node:fs",
+        "node:path",
+        "node:os",
+        "node:util",
+        "node:child_process",
+        "node:net",
+        "node:tty",
+        "node:stream",
+        "node:events",
+        "node:readline",
+        "node:module",
+        "node:url",
+        "node:https"
+      ]
     }
   ],
   "keywords": [
@@ -201,6 +283,7 @@
     "@types/node": "^25.2.0",
     "@vitest/coverage-v8": "^4.0.18",
     "husky": "^9.0.0",
+    "madge": "^8.0.0",
     "size-limit": "^12.0.0",
     "tsup": "^8.5.1",
     "tsx": "^4.21.0",
